Crystal tool: Try to figure out main project file.
This do a best effort solution for the case where a project has a root
file that require all other files, and then a macro is expanded in a
project file that make use of something that is defined in a different
file.

Is it a perfect solution? No, especially not for app projects having
multiple compile targets. But it works for libraries created with
"crystal init lib".